refactor: extract data models into models/ and core logic into simple/
