+++ src/pystormtracker/__init__.py
@@ -1,4 +1,4 @@
-from .detector import Center, RectGrid
-from .linker import Tracks
+from .models import Center, Grid, Tracks
+from .simple import SimpleDetector, SimpleLinker
 
-__all__ = ["Center", "RectGrid", "Tracks"]
+__all__ = ["Center", "Grid", "SimpleDetector", "SimpleLinker", "Tracks"]
