+++ src/pystormtracker/simple/__init__.py
@@ -1,0 +1,4 @@
+from .detector import SimpleDetector
+from .linker import SimpleLinker
+
+__all__ = ["SimpleDetector", "SimpleLinker"]
